Bulk transaction uploads

+++ src/components/ui/dashboard/TransactionsView.tsx
@@ -2,13 +2,9 @@ import { useState, useEffect } from "react";
 import Card from "@/components/ui/Card";
 import Modal from "@/components/ui/Modal";
 import api from "@/utils/api";
-import { IUser } from "@/components/context/UserContext";
 import TextInput from "@/components/ui/TextInput";
-import { IconPencil, IconTrash, IconPlus } from "@tabler/icons-react";
-
-interface TransactionsViewProps {
-  user: IUser;
-}
+import { IconPencil, IconTrash, IconPlus, IconUpload } from "@tabler/icons-react";
+import Papa from "papaparse"; // Import PapaParse for CSV parsing
 
 interface Transaction {
   _id: string;
@@ -16,20 +12,23 @@ interface Transaction {
   tags: string[];
   name: string;
   description: string;
-  createdAt: string;
-  updatedAt: string;
+  date: string;
 }
 
-export default function TransactionsView({ user }: TransactionsViewProps) {
+export default function TransactionsView() {
   const [isModalOpen, setIsModalOpen] = useState(false);
+  const [isBulkUploadModalOpen, setIsBulkUploadModalOpen] = useState(false);
   const [transactions, setTransactions] = useState<Transaction[]>([]);
   const [editTransaction, setEditTransaction] = useState<Transaction | null>(null);
+  const [isConfirmModalOpen, setIsConfirmModalOpen] = useState(false);
+  const [transactionToDelete, setTransactionToDelete] = useState<string | null>(null);
 
   // Individual state variables for each input
   const [value, setValue] = useState("0.0");
   const [tags, setTags] = useState("");
   const [name, setName] = useState("");
   const [description, setDescription] = useState("");
+  const [dateTime, setDateTime] = useState(new Date().toISOString().slice(0, 16)); // Default to now
 
   useEffect(() => {
     const fetchTransactions = async () => {
@@ -47,14 +46,15 @@ export default function TransactionsView({ user }: TransactionsViewProps) {
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
 
-    const tagsArray = tags.split(" ").map((tag) => tag.trim());
+    // Remove trailing and leading spaces from tags and split by space
+    const tagsArray = tags.trim().split(" ").map((tag) => tag.trim());
     // Prepare payload from individual states
     const payload = {
-      _id: editTransaction?._id,
       value: parseFloat(value) || 0,
       tags: tagsArray,
       name,
       description,
+      date: new Date(dateTime),
     };
 
     try {
@@ -78,6 +78,7 @@ export default function TransactionsView({ user }: TransactionsViewProps) {
       setTags("");
       setName("");
       setDescription("");
+      setDateTime(new Date().toISOString().slice(0, 16)); // Reset to now
 
       // Fetch the updated transactions
       const updatedTransactions = await api.get("/api/transaction/get");
@@ -93,21 +94,66 @@ export default function TransactionsView({ user }: TransactionsViewProps) {
     setTags(transaction.tags.join(" "));
     setName(transaction.name);
     setDescription(transaction.description);
+    setDateTime(new Date(transaction.date).toISOString().slice(0, 16));
     setIsModalOpen(true);
   };
 
-  const handleDelete = async (transactionId: string) => {
+  const handleDelete = async () => {
+    if (!transactionToDelete) return;
+
     try {
       const response = await api.delete(`/api/transaction/delete`, {
-        data: { transactionId },
+        data: { transactionId: transactionToDelete },
       });
       if (response.status === 200) {
         // Fetch the updated transactions
         const updatedTransactions = await api.get("/api/transaction/get");
         setTransactions(updatedTransactions.data);
+        setIsConfirmModalOpen(false);
+        setTransactionToDelete(null);
       }
     } catch (error) {
       console.error("Failed to delete transaction", error);
+    }
+  };
+
+  const confirmDelete = (transactionId: string) => {
+    setTransactionToDelete(transactionId);
+    setIsConfirmModalOpen(true);
+  };
+
+  const handleBulkUpload = async (e: React.FormEvent) => {
+    e.preventDefault();
+    const fileInput = document.getElementById("csvFileInput") as HTMLInputElement;
+    if (fileInput.files && fileInput.files[0]) {
+      Papa.parse(fileInput.files[0], {
+        header: true,
+        complete: async (results) => {
+          try {
+
+            // Foreach transaction in the CSV, parse and format the data
+            const finalTransactions = results.data.map((transaction: any) => {
+              return {
+                value: parseFloat(transaction.value) || 0,
+                tags: transaction.tags.split(",").map((tag: string) => tag.trim()),
+                name: transaction.name,
+                description: transaction.description,
+                date: new Date(transaction.date),
+              };
+            });
+
+            const response = await api.post("/api/transaction/bulk-add", finalTransactions);
+            if (response.status === 201) {
+              setIsBulkUploadModalOpen(false);
+              // Fetch the updated transactions
+              const updatedTransactions = await api.get("/api/transaction/get");
+              setTransactions(updatedTransactions.data);
+            }
+          } catch (error) {
+            console.error("Failed to bulk upload transactions", error);
+          }
+        },
+      });
     }
   };
 
@@ -115,12 +161,20 @@ export default function TransactionsView({ user }: TransactionsViewProps) {
     <div>
       <Card className="relative">
         <h1 className="text-2xl font-bold text-text">Transactions</h1>
-        <button
-          className="transition-all sm:text-lg text-sm absolute top-0 right-0 mt-4 mr-4 sm:px-2 sm:py-1 px-1 py-1 bg-primary text-white rounded flex items-center"
-          onClick={() => setIsModalOpen(true)}
-        >
-          <IconPlus className="" />
-        </button>
+        <div className="absolute top-0 right-0 mt-4 mr-4 flex space-x-2">
+          <button
+            className="transition-all sm:text-lg text-sm px-2 py-1 bg-primary text-white rounded flex items-center"
+            onClick={() => setIsModalOpen(true)}
+          >
+            <IconPlus className="mr-1" /> Add Transaction
+          </button>
+          <button
+            className="transition-all sm:text-lg text-sm px-2 py-1 bg-primary text-white rounded flex items-center"
+            onClick={() => setIsBulkUploadModalOpen(true)}
+          >
+            <IconUpload className="mr-1" /> Bulk Upload
+          </button>
+        </div>
 
         {/* Table Wrapper for Responsiveness */}
         <div className="overflow-x-auto">
@@ -138,7 +192,7 @@ export default function TransactionsView({ user }: TransactionsViewProps) {
             <tbody>
               {transactions.map((transaction: Transaction, index: number) => (
                 <tr
-                  key={transaction._id}
+                  key={index}
                   className={index < transactions.length - 1 ? "border-b border-backgroundGrayLight" : ""}
                 >
                   <td className="px-4 py-2">{transaction.name}</td>
@@ -161,16 +215,14 @@ export default function TransactionsView({ user }: TransactionsViewProps) {
                   </td>
                   <td className="px-4 py-2">{transaction.description}</td>
                   <td className="px-4 py-2 text-sm md:text-base">
-                    {new Date(transaction.createdAt).toLocaleString().length > 20
-                      ? new Date(transaction.createdAt).toLocaleTimeString([], { hour: '2-digit', minute: '2-digit', hour12: true })
-                      : new Date(transaction.createdAt).toLocaleString()}
+                    {new Date(transaction.date).toLocaleString()}
                   </td>
 
                   <td className="px-4 py-2 gap-2">
                     <button onClick={() => handleEdit(transaction)}>
                       <IconPencil />
                     </button>
-                    <button onClick={() => handleDelete(transaction._id)}>
+                    <button onClick={() => confirmDelete(transaction._id)}>
                       <IconTrash />
                     </button>
                   </td>
@@ -230,6 +282,17 @@ export default function TransactionsView({ user }: TransactionsViewProps) {
               />
             </label>
           </div>
+          <div>
+            <label>
+              Date and Time:
+              <TextInput
+                type="datetime-local"
+                value={dateTime}
+                onChange={(e) => setDateTime(e.target.value)}
+                required
+              />
+            </label>
+          </div>
           <button
             type="submit"
             className="w-full mt-4 px-4 py-2 bg-backgroundGray rounded"
@@ -237,6 +300,51 @@ export default function TransactionsView({ user }: TransactionsViewProps) {
             {editTransaction ? "Update Transaction" : "Create Transaction"}
           </button>
         </form>
+      </Modal>
+
+      <Modal open={isBulkUploadModalOpen} onClose={() => setIsBulkUploadModalOpen(false)}>
+        <form onSubmit={handleBulkUpload}>
+          <h1 className="text-2xl font-bold text-text">Bulk Upload</h1>
+          <div>
+            <label>
+              Upload CSV:
+              <input
+                id="csvFileInput"
+                type="file"
+                accept=".csv"
+                className="block w-full text-sm text-text border border-backgroundGrayLight rounded-lg cursor-pointer focus:outline-none"
+                required
+              />
+            </label>
+          </div>
+          <button
+            type="submit"
+            className="w-full mt-4 px-4 py-2 bg-backgroundGray rounded"
+          >
+            Upload
+          </button>
+        </form>
+      </Modal>
+
+      <Modal open={isConfirmModalOpen} onClose={() => setIsConfirmModalOpen(false)}>
+        <div className="p-4">
+          <h2 className="text-xl font-bold text-text">Are you sure?</h2>
+          <p className="text-text">Do you really want to delete this transaction? This process cannot be undone.</p>
+          <div className="flex justify-end mt-4">
+            <button
+              className="mr-2 px-4 py-2 bg-gray-300 text-gray-800 rounded"
+              onClick={() => setIsConfirmModalOpen(false)}
+            >
+              Cancel
+            </button>
+            <button
+              className="px-4 py-2 bg-red-500 text-white rounded"
+              onClick={handleDelete}
+            >
+              Delete
+            </button>
+          </div>
+        </div>
       </Modal>
     </div>
   );
